working with waltz
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@3495 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-compiler/src/test/java/org/drools/integrationtests/waltz/WaltzUtil.java
@@ -30,7 +30,7 @@ public class WaltzUtil {
 		return val % MOD_NUM;
 	}
 
-	private static int get_x(double val) {
+	private static int get_x(int val) {
 		return (int) (val / MOD_NUM);
 	}
